refactor(streamproxy): split streamproxy.go to mirror the test layout
Move-only split (catalog #18): urlheal, icy, tlsclock, bitrate, failures,
telemetry; the serve path stays in streamproxy.go.
