+++ internal/streamproxy/streamproxy.go
@@ -20,17 +20,13 @@ import (
 	"bufio"
 	"context"
 	"crypto/tls"
-	"crypto/x509"
 	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"log/slog"
 	"net"
 	"net/http"
-	"net/url"
-	"regexp"
 	"strconv"
 	"strings"
 	"sync"
@@ -39,211 +35,6 @@ import (
 	"github.com/JRpersonal/streborn/internal/netutil"
 	"github.com/JRpersonal/streborn/internal/presets"
 )
-
-// failLogSuppressWindow is the minimum spacing between identical
-// "upstream fail" warnings for the same URL. Bose's UPnP player
-// re-hits the proxy when a station is unreachable, so without this
-// the agent log fills with the same NXDOMAIN line several times a
-// minute for a single dead preset.
-const failLogSuppressWindow = 30 * time.Second
-
-// safeHTTPURL and the dial guard live in internal/netutil so the upnp playlist
-// fetcher shares the exact same SSRF policy. streamOne is reachable via handle()
-// with a URL straight from the preset store (CodeQL flagged that outbound Do()),
-// so the scheme gate stays mandatory here too.
-func safeHTTPURL(raw string) error { return netutil.SafeHTTPURL(raw) }
-
-// isHLSorDASHURL reports whether a URL points at an HLS (.m3u8) or DASH (.mpd)
-// playlist rather than a continuous raw stream. These are segment playlists,
-// not endless byte streams, and Bose's player cannot consume them; the proxy
-// would otherwise fetch the short playlist, hit EOF, and reconnect-loop forever
-// (BBC Radio 4 and the other BBC HLS-only stations). Until the
-// agent grows a real HLS/DASH remuxer, we detect these and report the stream as
-// not playable instead of looping. The query string is ignored so a tokenised
-// ".m3u8?..." still matches.
-func isHLSorDASHURL(raw string) bool {
-	return isHLSURL(raw) || isDASHURL(raw)
-}
-
-// isHLSURL reports whether a URL points at an HLS (.m3u8) playlist. STR follows
-// these and demuxes their segments (see serveHLS), so they are now playable.
-func isHLSURL(raw string) bool {
-	u, err := url.Parse(raw)
-	if err != nil {
-		return false
-	}
-	return strings.HasSuffix(strings.ToLower(u.Path), ".m3u8")
-}
-
-// isDASHURL reports whether a URL points at a DASH (.mpd) manifest. DASH is not
-// supported yet and is still refused.
-func isDASHURL(raw string) bool {
-	u, err := url.Parse(raw)
-	if err != nil {
-		return false
-	}
-	return strings.HasSuffix(strings.ToLower(u.Path), ".mpd")
-}
-
-// unwrapSelfProxy unwinds a stored stream URL that points back at the agent's own
-// stream proxy, e.g. "http://127.0.0.1:8888/stream/raw?u=<base64 real URL>". This
-// happens when a preset is saved from the box's now-playing location while a
-// radio station is playing THROUGH the proxy: the saved URL is the proxy wrapper,
-// not the real upstream (regression since v0.7.16, where ad-hoc radio routes via
-// /stream/raw). Recalling such a preset makes the proxy fetch its own loopback
-// URL, which the SSRF dial guard blocks, so the box gets nothing (INVALID_SOURCE /
-// AUDIO_ERROR_BAD_URL). Decoding the inner u recovers the real URL and heals the
-// preset in place, with no re-save needed. Loops in case of multiple wraps;
-// returns raw unchanged when it is not a self-proxy URL.
-func unwrapSelfProxy(raw string) string {
-	for i := 0; i < 5; i++ {
-		u, err := url.Parse(raw)
-		if err != nil || !strings.EqualFold(u.Path, "/stream/raw") {
-			return raw
-		}
-		enc := u.Query().Get("u")
-		if enc == "" {
-			return raw
-		}
-		dec, err := base64.RawURLEncoding.DecodeString(enc)
-		if err != nil {
-			if d2, e2 := base64.StdEncoding.DecodeString(enc); e2 == nil {
-				dec = d2
-			} else {
-				return raw
-			}
-		}
-		inner := string(dec)
-		if !strings.HasPrefix(inner, "http://") && !strings.HasPrefix(inner, "https://") {
-			return raw
-		}
-		raw = inner
-	}
-	return raw
-}
-
-// selfProxySlotRe matches the agent's own per-slot proxy path (/stream/1..6).
-var selfProxySlotRe = regexp.MustCompile(`^/stream/([1-6])$`)
-
-// selfProxySlotRef reports whether raw points at this agent's own
-// /stream/<slot> proxy (the box-visible preset location, never a valid
-// station origin) and which slot it references. Mirrors webui's save gate.
-func selfProxySlotRef(raw string) (int, bool) {
-	u, err := url.Parse(raw)
-	if err != nil {
-		return 0, false
-	}
-	m := selfProxySlotRe.FindStringSubmatch(u.Path)
-	if m == nil {
-		return 0, false
-	}
-	host, port := u.Hostname(), u.Port()
-	if port != "8888" && port != "17008" && host != "127.0.0.1" && host != "localhost" && host != "::1" {
-		return 0, false
-	}
-	n, _ := strconv.Atoi(m[1])
-	return n, true
-}
-
-// resolvePresetURL heals a stored preset URL before serving. Two poisoned
-// forms exist in the field (#252): the /stream/raw?u=<base64> self-wrap
-// (decoded by unwrapSelfProxy) and the bare /stream/<n> slot form an older
-// same-slot save left behind. For the slot form, n != slot dereferences the
-// referenced slot's stored origin (loop-guarded); n == slot is unrecoverable -
-// the origin URL is gone - and logs a distinct WARN naming the remedy instead
-// of the generic SSRF dial error the box otherwise trips.
-func (s *Server) resolvePresetURL(slot int, raw string) string {
-	raw = unwrapSelfProxy(raw)
-	seen := map[int]bool{slot: true}
-	for i := 0; i < 6; i++ {
-		ref, self := selfProxySlotRef(raw)
-		if !self {
-			return raw
-		}
-		if seen[ref] {
-			s.logger.Warn("stream proxy: preset stores its own proxy URL, origin lost - re-save the station in the app (#252)",
-				"slot", slot, "ref", ref)
-			return raw
-		}
-		seen[ref] = true
-		p, ok := s.store.Get(ref)
-		if !ok || p.StreamURL == "" {
-			s.logger.Warn("stream proxy: preset references another slot's proxy URL but that slot is empty - re-save the station in the app (#252)",
-				"slot", slot, "ref", ref)
-			return raw
-		}
-		raw = unwrapSelfProxy(p.StreamURL)
-	}
-	return raw
-}
-
-// isHLSorDASHContentType catches HLS/DASH responses whose URL does not carry a
-// telltale suffix, by their MIME type (application/vnd.apple.mpegurl,
-// application/x-mpegURL, audio/mpegurl, application/dash+xml).
-func isHLSorDASHContentType(ct string) bool {
-	ct = strings.ToLower(ct)
-	return strings.Contains(ct, "mpegurl") || strings.Contains(ct, "dash+xml")
-}
-
-// errPlaylistIsHLS signals that a response STR started to stream as a raw byte
-// stream turned out to be an HLS playlist (recognised by its body, not a .m3u8
-// URL suffix). The caller re-runs the request through serveHLS, which demuxes
-// the segments into one continuous stream. Returned only before any audio (or
-// response headers) have been written, so switching paths is safe.
-var errPlaylistIsHLS = errors.New("streamproxy: response is an HLS playlist")
-
-// hasHTTPScheme reports whether s begins with an http(s) scheme.
-func hasHTTPScheme(s string) bool {
-	return strings.HasPrefix(s, "http://") || strings.HasPrefix(s, "https://")
-}
-
-// firstMediaURLFromPlaylist extracts the first playable stream URL from a plain
-// M3U or PLS pointer playlist (a short text file that just lists the real stream
-// URL, as Icecast/Shoutcast directory stations and the Absolut* stations serve
-// under an audio/x-mpegurl MIME even when the URL has no .m3u suffix, #252).
-// M3U: the first non-comment line that is a URL. PLS: the first FileN=URL entry.
-// A relative M3U entry is resolved against baseURL. Returns "" when the body
-// carries no stream URL (e.g. it is actually an HLS segment playlist, which the
-// caller detects separately via the #EXT-X- markers).
-func firstMediaURLFromPlaylist(body, baseURL string) string {
-	base, _ := url.Parse(baseURL)
-	for _, line := range strings.Split(body, "\n") {
-		line = strings.TrimSpace(line)
-		if line == "" || strings.HasPrefix(line, "#") {
-			// blank line or M3U comment/directive (#EXTM3U, #EXTINF, #EXT-X-*).
-			continue
-		}
-		lower := strings.ToLower(line)
-		// PLS entry: FileN=URL (key is case-insensitive, N is a number).
-		if eq := strings.IndexByte(line, '='); eq > 0 && strings.HasPrefix(lower, "file") {
-			if cand := strings.TrimSpace(line[eq+1:]); hasHTTPScheme(cand) {
-				return cand
-			}
-			continue
-		}
-		// Absolute M3U entry.
-		if hasHTTPScheme(line) {
-			return line
-		}
-		// PLS metadata / section headers ([playlist], NumberOfEntries=1,
-		// Title1=..., Version=2) are not stream URLs.
-		if strings.ContainsAny(line, "=[]") {
-			continue
-		}
-		// A relative URI in an M3U — resolve it against the playlist URL.
-		if base != nil {
-			if ref, err := url.Parse(line); err == nil {
-				return base.ResolveReference(ref).String()
-			}
-		}
-	}
-	return ""
-}
-
-// hlsNotPlayableMsg is the user-facing reason returned to the box for an HLS/
-// DASH stream. Kept short; the box shows a not-playable state rather than a
-// silent reconnect storm.
-const hlsNotPlayableMsg = "HLS/DASH streams are not supported yet"
 
 type Server struct {
 	store  *presets.Store
@@ -344,52 +135,6 @@ type Server struct {
 	lastByteToBox time.Time
 }
 
-// SetOnTitle registers a callback invoked when the live ICY StreamTitle of
-// the proxied stream changes to a non-empty value. Set once at wiring time.
-func (s *Server) SetOnTitle(fn func(title string)) { s.onTitle = fn }
-
-// CurrentTitle returns the live ICY StreamTitle of the stream being proxied
-// right now, or "" if the station sends no metadata or none has arrived yet.
-func (s *Server) CurrentTitle() string {
-	s.titleMu.Lock()
-	defer s.titleMu.Unlock()
-	return s.curTitle
-}
-
-// setTitle records a freshly parsed StreamTitle for url and fires onTitle when
-// it changed to a new non-empty value. Empty titles (StreamTitle=”) clear the
-// current title but never fire the push, so a station that briefly sends an
-// empty title does not blank the box display with a spurious update.
-func (s *Server) setTitle(url, title string) {
-	title = strings.TrimRight(title, "\x00")
-	title = strings.TrimSpace(title)
-	s.titleMu.Lock()
-	changed := title != s.curTitle || url != s.curTitleURL
-	s.curTitle = title
-	s.curTitleURL = url
-	fire := changed && title != "" && s.onTitle != nil
-	cb := s.onTitle
-	s.titleMu.Unlock()
-	if changed {
-		s.logger.Info("stream proxy ICY title", "title", title)
-	}
-	if fire {
-		cb(title)
-	}
-}
-
-// clearTitleForNewURL drops a stale title when the proxied stream changes, so
-// the brief window before the new station's first metadata block does not show
-// the old station's track. A reconnect to the same url keeps the title.
-func (s *Server) clearTitleForNewURL(url string) {
-	s.titleMu.Lock()
-	if url != s.curTitleURL {
-		s.curTitle = ""
-		s.curTitleURL = url
-	}
-	s.titleMu.Unlock()
-}
-
 // SetOnDisconnect registers a callback invoked whenever the box closes a
 // proxied stream (raw or slot). Set once at wiring time.
 func (s *Server) SetOnDisconnect(fn func(upstreamErr error)) { s.onDisconnect = fn }
@@ -462,695 +207,12 @@ func New(store *presets.Store, logger *slog.Logger) *Server {
 	}
 }
 
-// icyConn wraps a net.Conn so a legacy SHOUTcast "ICY 200 OK" response line is
-// rewritten to "HTTP/1.0 200 OK" on the first read, letting Go's net/http parse
-// the response instead of rejecting it. All bytes after the status line (headers,
-// the ICY-interleaved audio) pass through unchanged. Only the very first line is
-// inspected; a normal "HTTP/1.x ..." response is left exactly as received.
-type icyConn struct {
-	net.Conn
-	br        *bufio.Reader
-	inspected bool
-	prefix    []byte // rewritten status-line bytes not yet handed to the caller
-}
-
-func (c *icyConn) Read(p []byte) (int, error) {
-	if !c.inspected {
-		c.inspected = true
-		// Peek only as far as the protocol token; blocks until the response
-		// arrives, which is exactly when http.Transport issues its first read.
-		if head, err := c.br.Peek(4); err == nil && string(head[:3]) == "ICY" && (head[3] == ' ' || head[3] == '\t') {
-			if line, err := c.br.ReadString('\n'); err == nil {
-				// "ICY 200 OK\r\n" -> "HTTP/1.0 200 OK\r\n" (keep the rest verbatim).
-				c.prefix = append([]byte("HTTP/1.0"), line[3:]...)
-			} else {
-				// No full line yet: hand back what we consumed unchanged so we
-				// never lose bytes; the transport will keep reading.
-				c.prefix = []byte(line)
-			}
-		}
-	}
-	if len(c.prefix) > 0 {
-		n := copy(p, c.prefix)
-		c.prefix = c.prefix[n:]
-		return n, nil
-	}
-	return c.br.Read(p)
-}
-
-// minPlausibleClock is a lower bound on a trustworthy wall clock. STR shipped in
-// 2026, so a box reporting a time before this has an unset clock. SoundTouch
-// speakers have no battery-backed RTC: after a cold boot, before NTP has synced
-// (or when NTP is blocked), the clock lands in the firmware's build epoch, often
-// mid-2015. See #296 and docs/FIRMWARE-NOTES.md.
-var minPlausibleClock = time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC)
-
-// clockUntrustworthy reports whether the local wall clock is implausibly old,
-// which would make certificate time-validity checks fail spuriously. It reads
-// time.Now() live, so strict verification resumes automatically the moment NTP
-// corrects the clock.
-func clockUntrustworthy() bool { return time.Now().Before(minPlausibleClock) }
-
-// clockTolerantTLSConfig builds the per-connection TLS config the proxy uses for
-// an upstream HTTPS radio stream dialed at host (a hostname, or a bare-IP
-// literal). It always verifies the certificate chain to the system roots and the
-// host, but tolerates a wrong box clock: when the local clock is implausibly
-// old, a chain that is valid except for the certificate's time window is still
-// accepted. Without this, a speaker whose clock reset to 2015 rejected every
-// HTTPS station as "certificate is not yet valid" (#296: Virgin Radio and other
-// HTTPS streams would not play, while plain-HTTP BBC streams still did).
-//
-// Only the expiry/not-yet-valid window is relaxed, and only while the clock is
-// untrustworthy; chain-to-root and host are always enforced, so this does not
-// weaken MITM protection. host is passed explicitly (not read from
-// ConnectionState.ServerName) so a bare-IP upstream, which sends no SNI, is
-// still verified against the certificate's IP SANs.
-func clockTolerantTLSConfig(host string) *tls.Config {
-	return &tls.Config{
-		ServerName: host, // SNI for a hostname; ignored on the wire for an IP literal
-		// h2 stays available (DialTLSContext otherwise defaults to HTTP/1.1 only).
-		NextProtos: []string{"h2", "http/1.1"},
-		// The default verifier is disabled so the time window can be relaxed in
-		// VerifyConnection; chain and host are still checked there manually.
-		InsecureSkipVerify: true, //nolint:gosec // VerifyConnection re-implements chain+host verification below.
-		VerifyConnection: func(cs tls.ConnectionState) error {
-			roots, _ := x509.SystemCertPool()
-			return verifyChainClockTolerant(cs.PeerCertificates, host, roots, time.Now(), clockUntrustworthy())
-		},
-	}
-}
-
-// verifyChainClockTolerant verifies leaf (certs[0]) against roots and host at
-// time now. host is matched against the certificate's DNS names, or, for an IP
-// literal, its IP SANs. If verification fails solely because the certificate is
-// outside its time window and clockBad is true, it retries with the time check
-// pinned to the leaf's NotBefore (so the window always passes) while still
-// requiring a valid chain and host. It is split out from clockTolerantTLSConfig
-// so the policy can be unit-tested without a live TLS handshake.
-func verifyChainClockTolerant(certs []*x509.Certificate, host string, roots *x509.CertPool, now time.Time, clockBad bool) error {
-	if len(certs) == 0 {
-		return errors.New("tls: no peer certificates")
-	}
-	// Require a host: x509.Verify silently skips host checking when DNSName is
-	// empty, so an empty host here would drop the host guard entirely. The
-	// caller derives it from the dial address, so an empty value means something
-	// is wrong and we must reject rather than relax.
-	if host == "" {
-		return errors.New("tls: empty host")
-	}
-	inter := x509.NewCertPool()
-	for _, c := range certs[1:] {
-		inter.AddCert(c)
-	}
-	leaf := certs[0]
-	opts := x509.VerifyOptions{DNSName: host, Roots: roots, Intermediates: inter, CurrentTime: now}
-	if _, err := leaf.Verify(opts); err == nil {
-		return nil
-	} else {
-		// Relax the time window only when the clock is untrustworthy and the
-		// failure was a time-validity problem (x509.Expired covers both
-		// not-yet-valid and expired). Any other failure (unknown authority,
-		// hostname mismatch) still rejects the connection.
-		var invalid x509.CertificateInvalidError
-		if clockBad && errors.As(err, &invalid) && invalid.Reason == x509.Expired {
-			opts.CurrentTime = leaf.NotBefore
-			if _, err2 := leaf.Verify(opts); err2 == nil {
-				return nil
-			}
-		}
-		return err
-	}
-}
-
-// CurrentBitrate returns the detected bitrate (kbps) of the stream being
-// proxied right now, or 0 if unknown.
-func (s *Server) CurrentBitrate() int {
-	s.brMu.Lock()
-	defer s.brMu.Unlock()
-	return s.curBitrate
-}
-
-// beginStream marks url as the stream now playing and seeds curBitrate:
-// the icy-br value if the station sent one, else a value already measured
-// for this URL in a previous play (so a reconnect or a re-play does not
-// briefly show "-"), else 0 (unknown, to be measured). Crucially it always
-// updates curURL so switching from a station that had a bitrate to one that
-// does not clears the stale number instead of leaving the old one showing.
-func (s *Server) beginStream(url string, icyBr int) (known bool) {
-	s.brMu.Lock()
-	defer s.brMu.Unlock()
-	s.curURL = url
-	if icyBr > 0 {
-		s.curBitrate = icyBr
-		s.measuredBr[url] = icyBr
-		return true
-	}
-	if br, ok := s.measuredBr[url]; ok {
-		s.curBitrate = br
-		return true
-	}
-	s.curBitrate = 0
-	return false
-}
-
-// rememberBitrate stores a throughput-measured bitrate for url and makes it
-// the current value, so internal reconnects reuse it rather than measuring
-// a fresh (and slightly different) number on every UI poll. The map is
-// capped so the ad-hoc search-play path cannot grow it without bound.
-func (s *Server) rememberBitrate(url string, br int) {
-	if br <= 0 {
-		return
-	}
-	s.brMu.Lock()
-	defer s.brMu.Unlock()
-	if len(s.measuredBr) > 64 {
-		s.measuredBr = make(map[string]int)
-	}
-	s.measuredBr[url] = br
-	s.curURL = url
-	s.curBitrate = br
-}
-
-// icyBitrate pulls the real stream bitrate (kbps) from the Icecast/
-// Shoutcast response headers, 0 if none present. icy-br is sometimes a
-// comma list ("128,128"); the first value is used.
-func icyBitrate(h http.Header) int {
-	for _, k := range []string{"icy-br", "ice-bitrate", "x-audiocast-bitrate"} {
-		v := h.Get(k)
-		if v == "" {
-			continue
-		}
-		if i := strings.IndexByte(v, ','); i >= 0 {
-			v = v[:i]
-		}
-		if n, err := strconv.Atoi(strings.TrimSpace(v)); err == nil && n > 0 {
-			return n
-		}
-	}
-	return 0
-}
-
-// icyMetaint returns the byte spacing between interleaved ICY metadata blocks
-// from the upstream icy-metaint response header, or 0 if the station sends no
-// metadata. With a non-zero value the stream is: metaint audio bytes, then one
-// length byte L, then L*16 bytes of metadata, repeating.
-func icyMetaint(h http.Header) int {
-	v := h.Get("icy-metaint")
-	if v == "" {
-		return 0
-	}
-	if n, err := strconv.Atoi(strings.TrimSpace(v)); err == nil && n > 0 {
-		return n
-	}
-	return 0
-}
-
-// parseStreamTitle pulls the track text out of an ICY metadata block, which
-// looks like `StreamTitle='Artist - Song';StreamUrl='...';` padded to a 16-byte
-// boundary with NULs. Returns ok=false when there is no StreamTitle field.
-func parseStreamTitle(meta string) (string, bool) {
-	const key = "StreamTitle='"
-	i := strings.Index(meta, key)
-	if i < 0 {
-		return "", false
-	}
-	rest := meta[i+len(key):]
-	// Closing delimiter is `';`; fall back to a lone quote if the station omits
-	// the semicolon, and to the whole remainder (NUL-trimmed) as a last resort.
-	if j := strings.Index(rest, "';"); j >= 0 {
-		return rest[:j], true
-	}
-	if j := strings.IndexByte(rest, '\''); j >= 0 {
-		return rest[:j], true
-	}
-	return strings.TrimRight(rest, "\x00"), true
-}
-
-// icyReader wraps an upstream stream that carries interleaved ICY metadata and
-// presents only the audio bytes to the caller. Each metadata block is handed
-// to onMeta as it is read, so the proxy can extract StreamTitle without ever
-// forwarding the metadata (or the icy-metaint contract) to the box.
-type icyReader struct {
-	src     io.Reader
-	metaint int
-	remain  int // audio bytes left before the next metadata block
-	onMeta  func(meta string)
-}
-
-func newICYReader(src io.Reader, metaint int, onMeta func(meta string)) *icyReader {
-	return &icyReader{src: src, metaint: metaint, remain: metaint, onMeta: onMeta}
-}
-
-func (r *icyReader) Read(p []byte) (int, error) {
-	// At a metadata boundary: read the length byte and, if non-zero, the block.
-	if r.remain == 0 {
-		var lb [1]byte
-		if _, err := io.ReadFull(r.src, lb[:]); err != nil {
-			return 0, err
-		}
-		if mlen := int(lb[0]) * 16; mlen > 0 {
-			meta := make([]byte, mlen)
-			if _, err := io.ReadFull(r.src, meta); err != nil {
-				return 0, err
-			}
-			if r.onMeta != nil {
-				r.onMeta(string(meta))
-			}
-		}
-		r.remain = r.metaint
-	}
-	// Read at most up to the next metadata boundary so the length byte is never
-	// mistaken for audio.
-	n := len(p)
-	if n > r.remain {
-		n = r.remain
-	}
-	read, err := r.src.Read(p[:n])
-	r.remain -= read
-	return read, err
-}
-
-// standardBitrates are the common MP3/AAC stream rates a throughput
-// measurement is snapped to, so the displayed value is stable and
-// realistic instead of a jittery raw number.
-var standardBitrates = []int{32, 48, 64, 80, 96, 112, 128, 160, 192, 224, 256, 320, 384, 448, 512}
-
-// roundStandardBitrate snaps a throughput-derived kbps to the nearest
-// common audio stream rate. Steady-state throughput of a playing stream
-// sits very close to its real bitrate, so the nearest standard rate is the
-// honest value. Anything above the highest audio rate is still buffer-fill
-// burst, not a real bitrate, and returns 0 (shown as "-" rather than a
-// misleading number like 1310).
-func roundStandardBitrate(kbps int) int {
-	if kbps <= 0 || kbps > 600 {
-		return 0
-	}
-	best, bestDelta := 0, 1<<30
-	for _, std := range standardBitrates {
-		d := kbps - std
-		if d < 0 {
-			d = -d
-		}
-		if d < bestDelta {
-			bestDelta, best = d, std
-		}
-	}
-	return best
-}
-
-// shouldLogFail returns true if a fresh WARN line should be emitted
-// for this URL. Within failLogSuppressWindow of the previous emit it
-// returns false so the agent log does not repeat the same
-// "upstream fail" line every time Bose's UPnP player retries an
-// unreachable station.
-func (s *Server) shouldLogFail(url string) bool {
-	s.failMu.Lock()
-	defer s.failMu.Unlock()
-	now := time.Now()
-	if last, ok := s.lastFail[url]; ok && now.Sub(last) < failLogSuppressWindow {
-		return false
-	}
-	s.lastFail[url] = now
-	return true
-}
-
-// markByteDelivery stamps the moment audio last reached the box. Called from the
-// streamOne copy loop on every successful write; cheap (one guarded time.Now),
-// never logged per call.
-func (s *Server) markByteDelivery() {
-	s.gapMu.Lock()
-	s.lastByteToBox = time.Now()
-	s.gapMu.Unlock()
-}
-
-// resetAudioGap clears the last-byte stamp at the start of a fresh stream so a
-// stale gap from a previous station is not reported on the first reconnect.
-func (s *Server) resetAudioGap() {
-	s.gapMu.Lock()
-	s.lastByteToBox = time.Time{}
-	s.gapMu.Unlock()
-}
-
-// audioGap returns how long it has been since audio last reached the box, or 0
-// if no byte has been delivered yet (so the very first connect attempt is not
-// reported as a gap).
-func (s *Server) audioGap() time.Duration {
-	s.gapMu.Lock()
-	defer s.gapMu.Unlock()
-	if s.lastByteToBox.IsZero() {
-		return 0
-	}
-	return time.Since(s.lastByteToBox)
-}
-
-// upstreamStatusError carries the upstream HTTP status of a non-200 response so
-// the reconnect loops can tell a permanent client-side rejection (403 geo-block,
-// 404/410 gone) from a transient one (5xx) via errors.As, instead of parsing a
-// formatted string.
-type upstreamStatusError struct {
-	Code   int
-	Status string
-}
-
-func (e *upstreamStatusError) Error() string {
-	if e.Status != "" {
-		return "upstream status " + e.Status
-	}
-	return fmt.Sprintf("upstream status %d", e.Code)
-}
-
-// isPermanentUpstream reports whether err is an upstream failure that retrying
-// the SAME URL cannot fix: a client-side HTTP rejection (forbidden, not found,
-// gone, unavailable-for-legal-reasons) or an HLS/DASH playlist we cannot play.
-// For these the reconnect loop gives up immediately so the desktop app can fall
-// back to another radio-browser entry of the station within a second instead of
-// after a 30s retry storm against a URL that will never serve audio.
-func isPermanentUpstream(err error) bool {
-	var se *upstreamStatusError
-	if errors.As(err, &se) {
-		switch se.Code {
-		case http.StatusBadRequest, http.StatusUnauthorized, http.StatusForbidden,
-			http.StatusNotFound, http.StatusGone, http.StatusUnavailableForLegalReasons:
-			return true
-		}
-	}
-	return false
-}
-
-// streamFailure is the most recent terminal upstream error, surfaced via
-// /api/stream-status. url is the upstream stream URL (not the proxy wrapper) so
-// the app can confirm the error belongs to the station it just started.
-type streamFailure struct {
-	when   time.Time
-	code   int    // upstream HTTP status, 0 for a network-level failure
-	reason string // coarse class the UI maps to a message: blocked|gone|unavailable|unreachable|hls
-	url    string
-}
-
-// classifyFailure maps an upstream error to a coarse reason the desktop app
-// turns into a localized, human message. Network errors (DNS, refused, reset)
-// land in "unreachable"; HTTP statuses split into blocked/gone/unavailable.
-func classifyFailure(err error) (code int, reason string) {
-	var se *upstreamStatusError
-	if errors.As(err, &se) {
-		switch se.Code {
-		case http.StatusUnauthorized, http.StatusForbidden, http.StatusUnavailableForLegalReasons:
-			return se.Code, "blocked"
-		case http.StatusNotFound, http.StatusGone:
-			return se.Code, "gone"
-		default:
-			// Everything else (5xx, and any other non-200) is a transient
-			// "currently unavailable" from the station's side.
-			return se.Code, "unavailable"
-		}
-	}
-	if strings.Contains(strings.ToLower(errStr(err)), "hls") || strings.Contains(strings.ToLower(errStr(err)), "dash") {
-		return 0, "hls"
-	}
-	return 0, "unreachable"
-}
-
-// recordFailure stores the latest terminal upstream failure for url so the
-// desktop app can poll it and react (message + alternative-source fallback).
-func (s *Server) recordFailure(url string, err error) {
-	code, reason := classifyFailure(err)
-	s.errMu.Lock()
-	s.lastErr = streamFailure{when: time.Now(), code: code, reason: reason, url: url}
-	s.errMu.Unlock()
-}
-
-// boxHasOutbound reports whether the speaker itself can reach the public
-// internet, cached for a few seconds. It lets handleStreamStatus tell a single
-// unreachable station apart from the speaker being offline entirely (#375): a
-// box on a dead Wi-Fi fails EVERY station with a network-level error, which
-// otherwise reads as each station being individually broken. Dials a couple of
-// well-known hosts by NAME on :443 (so DNS, which radio also needs, is part of
-// the test) plus a raw resolver IP; any success means online.
-func (s *Server) boxHasOutbound() bool {
-	s.netMu.Lock()
-	if !s.netCheckAt.IsZero() && time.Since(s.netCheckAt) < 8*time.Second {
-		v := s.netOnline
-		s.netMu.Unlock()
-		return v
-	}
-	s.netMu.Unlock()
-	online := false
-	for _, h := range []string{"www.google.com:443", "www.cloudflare.com:443", "1.1.1.1:53"} {
-		c, err := net.DialTimeout("tcp", h, 2*time.Second)
-		if err == nil {
-			_ = c.Close()
-			online = true
-			break
-		}
-	}
-	s.netMu.Lock()
-	s.netOnline = online
-	s.netCheckAt = time.Now()
-	s.netMu.Unlock()
-	return online
-}
-
-// clearFailure drops any recorded failure for url once it streams successfully,
-// so a recovered station does not keep reporting a stale error to the app.
-func (s *Server) clearFailure(url string) {
-	s.errMu.Lock()
-	if s.lastErr.url == url {
-		s.lastErr = streamFailure{}
-	}
-	s.errMu.Unlock()
-}
-
-// Register registers /stream/<slot> as well as /stream/raw for ad-hoc URLs
-// (e.g. from the radio search) on the supplied mux.
-// noteFetchOpen records that the box opened a proxied stream just now and
-// marks the connection open; the returned func marks it closed and re-stamps
-// the activity, so "served until the connection dropped a moment ago" reads
-// as fresh. A bare open-time stamp was not enough: a radio stream holds ONE
-// long GET for its whole playback, so after 9m42s of flawless audio the
-// spontaneous-off recovery read lastFetch as ten minutes stale and stood
-// down, leaving the speaker dead after a firmware source-drop (#491
-// Cinemate, field bundle 2026-08-01).
-func (s *Server) noteFetchOpen() func() {
-	s.fetchMu.Lock()
-	s.lastFetch = time.Now()
-	s.openFetches++
-	s.fetchMu.Unlock()
-	return func() {
-		s.fetchMu.Lock()
-		s.lastFetch = time.Now()
-		s.openFetches--
-		s.fetchMu.Unlock()
-	}
-}
-
-// noteSlotFetch records that the box opened THIS slot's proxied stream. Called
-// only after the slot validated and the store had a playable preset, so a 404
-// or a foreign fetch can never stamp it. Paired with noteSlotFetchDone.
-func (s *Server) noteSlotFetch(slot int) {
-	if slot < 1 || slot > 6 {
-		return
-	}
-	s.fetchMu.Lock()
-	s.slotFetch[slot] = time.Now()
-	s.slotOpen[slot]++
-	s.fetchMu.Unlock()
-}
-
-// noteSlotFetchDone records that a slot connection closed, for the liveness
-// half of SlotPulledSince.
-func (s *Server) noteSlotFetchDone(slot int) {
-	if slot < 1 || slot > 6 {
-		return
-	}
-	s.fetchMu.Lock()
-	if s.slotOpen[slot] > 0 {
-		s.slotOpen[slot]--
-	}
-	s.slotFetchEnd[slot] = time.Now()
-	s.fetchMu.Unlock()
-}
-
-// LastFetchForSlot reports when the box last opened the given slot's proxied
-// stream (zero time = never, or slot out of range). The global LastActivity
-// stamp stays for the wedge detector, which deliberately counts any fetch.
-// No production caller since the recall verify moved to the liveness-aware
-// SlotPulledSince (a bare open-stamp certified failed recalls); kept as the
-// low-level accessor its tests and future diagnostics read the raw stamp
-// through.
-func (s *Server) LastFetchForSlot(slot int) time.Time {
-	if slot < 1 || slot > 6 {
-		return time.Time{}
-	}
-	s.fetchMu.Lock()
-	defer s.fetchMu.Unlock()
-	return s.slotFetch[slot]
-}
-
-// minSustainedFetch is how long a now-closed slot fetch must have served to
-// still count as "the box played this recall". The box's re-login source
-// bounce opens the stream for 36ms-2.4s and drops it; a genuine playback
-// session either stays open or lasted well past this.
-const minSustainedFetch = 3 * time.Second
-
-// SlotPulledSince reports whether the box is credibly playing this slot's
-// proxied stream for a recall anchored at t: a connection opened after t that
-// is still OPEN, or one that served at least minSustainedFetch before closing.
-// This is the hardware recall verify's success signal; "opened once since t"
-// alone certified dead recalls as healthy (#252 field bundles).
-func (s *Server) SlotPulledSince(slot int, t time.Time) bool {
-	if slot < 1 || slot > 6 {
-		return false
-	}
-	s.fetchMu.Lock()
-	defer s.fetchMu.Unlock()
-	start := s.slotFetch[slot]
-	if start.IsZero() || !start.After(t) {
-		return false
-	}
-	if s.slotOpen[slot] > 0 {
-		return true
-	}
-	return s.slotFetchEnd[slot].Sub(start) >= minSustainedFetch
-}
-
-// SlotFetchLive reports whether the box currently has an OPEN connection to this
-// slot's proxied stream. Unlike SlotPulledSince it makes no sustained-duration
-// judgement on a closed fetch: it is the "is the box pulling audio right now"
-// signal the recall verify uses to distinguish a genuine playback from a login
-// error's brief source bounce (which opens the stream, serves ~3s and closes).
-func (s *Server) SlotFetchLive(slot int) bool {
-	if slot < 1 || slot > 6 {
-		return false
-	}
-	s.fetchMu.Lock()
-	defer s.fetchMu.Unlock()
-	return s.slotOpen[slot] > 0
-}
-
-// SetBoxStateFn wires the speaker-side condition reporter for
-// /api/stream-status (see boxStateFn).
-func (s *Server) SetBoxStateFn(fn func() string) {
-	s.boxStateFn = fn
-}
-
-// LastActivity reports when the box was last SERVED by any proxied stream
-// (activity is "now" while a connection is open, the close moment after it
-// ends, the open moment otherwise) and when the last terminal upstream
-// failure happened (zero times = never). Consumed by the webui's wedge
-// detector, the recall inert-check and the spontaneous-off recovery.
-func (s *Server) LastActivity() (lastFetch, lastFailure time.Time) {
-	s.fetchMu.Lock()
-	lastFetch = s.lastFetch
-	if s.openFetches > 0 {
-		lastFetch = time.Now()
-	}
-	s.fetchMu.Unlock()
-	s.errMu.Lock()
-	lastFailure = s.lastErr.when
-	s.errMu.Unlock()
-	return lastFetch, lastFailure
-}
-
 func (s *Server) Register(mux *http.ServeMux) {
 	mux.HandleFunc("/stream/raw", s.handleRaw)
 	mux.HandleFunc("/stream/", s.handle)
 	mux.HandleFunc("/api/stream/bitrate", s.handleBitrate)
 	mux.HandleFunc("/api/stream/title", s.handleTitle)
 	mux.HandleFunc("/api/stream-status", s.handleStreamStatus)
-}
-
-// handleStreamStatus reports the most recent terminal upstream failure as JSON:
-//
-//	{"error":true,"status":403,"reason":"blocked","url":"http://...","ageMs":1200}
-//
-// or {"error":false} when the last start streamed fine or is too old to matter.
-// The desktop app polls this for a few seconds after starting a station: a fresh
-// error tells it which message to show and that it should try another
-// radio-browser entry of the same station. Only failures younger than
-// streamStatusTTL are reported so a long-past error never blocks a fresh play.
-func (s *Server) handleStreamStatus(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	// boxState surfaces a SPEAKER-side condition ("wedged", "login-error")
-	// that makes every station fail, so the app can say what is actually
-	// wrong instead of blaming the station and cycling radio-browser
-	// alternates ("Sender spielt nicht ... suche andere Quelle" while the box
-	// was rejecting sources as not-logged-in). Additive: absent when the box
-	// is fine, ignored by older frontends.
-	boxState := ""
-	if s.boxStateFn != nil {
-		boxState = s.boxStateFn()
-	}
-	s.errMu.Lock()
-	f := s.lastErr
-	s.errMu.Unlock()
-	if f.when.IsZero() || time.Since(f.when) > streamStatusTTL {
-		if boxState != "" {
-			fmt.Fprintf(w, `{"error":false,"boxState":%q}`, boxState)
-			return
-		}
-		fmt.Fprint(w, `{"error":false}`)
-		return
-	}
-	// A network-level "unreachable" is ambiguous: the one station may be down,
-	// or the speaker has no internet at all. When outbound is also down, report a
-	// distinct "offline" reason so the app can say "the speaker has no internet
-	// connection" and offer to re-run Wi-Fi setup, instead of blaming every
-	// station in turn (#375). Only checked on the "unreachable" catch-all, so a
-	// clear 403/blocked/gone is never masked and the probe stays off the hot path.
-	if f.reason == "unreachable" && !s.boxHasOutbound() {
-		f.reason = "offline"
-	}
-	body, err := json.Marshal(struct {
-		Error    bool   `json:"error"`
-		Status   int    `json:"status"`
-		Reason   string `json:"reason"`
-		URL      string `json:"url"`
-		AgeMs    int64  `json:"ageMs"`
-		BoxState string `json:"boxState,omitempty"`
-	}{
-		Error:    true,
-		Status:   f.code,
-		Reason:   f.reason,
-		URL:      f.url,
-		AgeMs:    time.Since(f.when).Milliseconds(),
-		BoxState: boxState,
-	})
-	if err != nil {
-		fmt.Fprint(w, `{"error":false}`)
-		return
-	}
-	w.Write(body)
-}
-
-// streamStatusTTL bounds how long a recorded upstream failure is reported. Long
-// enough for the app's post-play poll window, short enough that a stale error
-// from minutes ago never suppresses a fresh, healthy station.
-const streamStatusTTL = 20 * time.Second
-
-// handleTitle returns the live ICY StreamTitle of the stream currently
-// proxied, or "" when the station sends no metadata. Cheap: a single guarded
-// string read. The desktop app polls this on a slow cadence to show the live
-// radio track next to the station name, the same way it shows the bitrate.
-func (s *Server) handleTitle(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	writeJSONString(w, "title", s.CurrentTitle())
-}
-
-// writeJSONString emits {"<key>":"<value>"} with value JSON-escaped, so a
-// StreamTitle containing quotes or backslashes cannot break the response.
-func writeJSONString(w io.Writer, key, value string) {
-	esc := strings.NewReplacer(`\`, `\\`, `"`, `\"`, "\n", `\n`, "\r", `\r`, "\t", `\t`)
-	fmt.Fprintf(w, `{%q:"%s"}`, key, esc.Replace(value))
-}
-
-// handleBitrate returns the real bitrate (kbps) of the stream currently
-// proxied, detected from the upstream icy-br header or a throughput
-// sample. 0 means unknown. Cheap: a single guarded int read. The desktop
-// app fetches this once per station change (not on a timer) to keep box
-// load minimal.
-func (s *Server) handleBitrate(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	fmt.Fprintf(w, `{"bitrate":%d}`, s.CurrentBitrate())
 }
 
 // handleRaw streams an arbitrary URL through — used by the radio search
@@ -1621,13 +683,4 @@ func (s *Server) streamOneDepth(ctx context.Context, w http.ResponseWriter, r *h
 			return true, readErr
 		}
 	}
-}
-
-// errStr renders an error for a structured log field, empty when nil so
-// a clean stream end does not log a misleading "lastErr=<nil>".
-func errStr(err error) string {
-	if err == nil {
-		return ""
-	}
-	return err.Error()
 }
